backend/requirements.txt
# ========================
# Core Framework
# ========================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ========================
# Database & ORM
# ========================
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# ========================
# Authentication
# ========================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-oauth2-client>=0.3.0

# ========================
# Redis & Celery
# ========================
redis>=5.0.1
celery>=5.3.6

# ========================
# ML & NLP
# ========================
sentence-transformers>=2.5.1
torch>=2.1.0
transformers>=4.36.0
numpy>=1.26.3
scikit-learn>=1.4.0

# ========================
# Document Processing
# ========================
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23
pdf2image>=1.17.0
pytesseract>=0.3.10

# ========================
# Storage (MinIO/S3)
# ========================
boto3>=1.34.0
aiobotocore>=2.7.0

# ========================
# Configuration
# ========================
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ========================
# Utilities
# ========================
aiofiles>=23.2.1
python-dateutil>=2.8.2
tenacity>=8.2.3
tqdm>=4.66.1

# ========================
# Testing
# ========================
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# ========================
# Development
# ========================
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

